Fix the processing of JSON-RPC requests encoded using the GET method
